Initial commit.
To say this is the "initial commit" is a bit misleading. The functions in the file have been developed over several years, initially to solve some startup ready detection problems encountered by developers who were trying to port existing web apps into the original Intel XDK container and, eventually, the standard Cordova container. It is the first commit to this repository...